code refactor + added and fixed comments
