+++ rpc/src/encoding/mod.rs
@@ -1,6 +1,7 @@
 pub mod base_types;
 pub mod monitor;
 pub mod chain;
+pub mod conversions;
 pub mod context;
 
 #[cfg(test)]
